+++ app/src/main/java/org/schabi/newpipe/BaseActivity.kt
@@ -1,9 +1,9 @@
 /*
- * SPDX-FileCopyrightText: 2025-2026 NewPipe e.V. <https://newpipe-ev.de>
+ * SPDX-FileCopyrightText: 2015-2026 NewPipe contributors <https://newpipe.net>
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-package org.schabi.newpipe.ui
+package org.schabi.newpipe
 
 import android.graphics.Color
 import android.os.Build
@@ -17,7 +17,7 @@ import org.schabi.newpipe.ui.theme.AppTheme
 
 /**
  * Base activity for Compose-based screens. Provides edge-to-edge display and
- * wraps Compose content in [AppTheme].
+ * wraps Compose content in [org.schabi.newpipe.ui.theme.AppTheme].
  *
  * Subclasses should be annotated with `@AndroidEntryPoint` if they need Hilt injection.
  */
@@ -25,7 +25,7 @@ open class BaseActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         enableEdgeToEdge(
-            navigationBarStyle = SystemBarStyle.auto(Color.TRANSPARENT, Color.TRANSPARENT)
+            navigationBarStyle = SystemBarStyle.Companion.auto(Color.TRANSPARENT, Color.TRANSPARENT)
         )
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
             window.isNavigationBarContrastEnforced = false
@@ -33,9 +33,6 @@ open class BaseActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
     }
 
-    /**
-     * Sets the Compose content wrapped in [AppTheme]. Call this instead of [setContent] directly.
-     */
     fun composeSetContent(content: @Composable () -> Unit) {
         setContent {
             AppTheme(content = content)
